Bot_Response1.0/requirements.txt
google-generativeai
langchain-google-genai
streamlit
flask
python-dotenv